refactor(tests): split test_httpx into a httpx/ test package
tests/tools/recon/test_httpx.py (554 lines) split to mirror the
tools/recon/httpx/ source package: test_flags.py (TestAssembleFlags),
test_scanner.py (httpx_scan modes / evidence / probe_endpoints shim),
test_parser.py (the _build_tls_certificate / _parse_cert_datetime
helpers). Each well under 450; classes move verbatim.
